+++ src/api/product/content-types/product/lifecycles.js
@@ -1,8 +1,27 @@
-'use strict';
+const generateSlug = (text) => {
+  return text
+    .toString()
+    .toLowerCase()
+    .normalize('NFD')
+    .replace(/[\u0300-\u036f]/g, '')
+    .trim()
+    .replace(/\s+/g, '-')
+    .replace(/[^\w-]+/g, '')
+    .replace(/--+/g, '-');
+};
 
-/**
- * Read the documentation (https://strapi.io/documentation/developer-docs/latest/development/backend-customization.html#lifecycle-hooks)
- * to customize this model
- */
+module.exports = {
+  async beforeCreate(event) {
+    const { data } = event.params;
+    if (data.name) {
+      data.slug = data.slug || generateSlug(data.name);
+    }
+  },
 
-module.exports = {};
+  async beforeUpdate(event) {
+    const { data } = event.params;
+    if (data.name && !data.slug) {
+      data.slug = generateSlug(data.name);
+    }
+  },
+};