+++ src/api/user-credential/routes/1-user-credential-custom.js
@@ -1,0 +1,13 @@
+module.exports = {
+  routes: [
+    {
+      method: 'POST',
+      path: '/user-credentials/:documentId/add-products',
+      handler: 'api::user-credential.user-credential.addProducts',
+      config: {
+        auth: false,
+        policies: [],
+      },
+    },
+  ],
+};