+++ src/api/user-credential/services/user-credential.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * user-credential service
+ */
+
+const { createCoreService } = require('@strapi/strapi').factories;
+
+module.exports = createCoreService('api::user-credential.user-credential');