+++ CHANGELOG.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - 2026-04-17
+## [1.2.0] - 2026-04-24
 
 ## Added
 
 - **Apim Config:** Implement on-demand and periodic synchronization for APIM configurations
+- **Apim Config:** Implement credentials management for APIM configurations
+- **Users Credentials:** Implement credentials for apiKey and oAuth2 providers
+- **Products:** Add support for product-level credentials in APIM configurations
+- **Library Api:** Add external service id to library API for better integration with external services
 
 
 ## [1.1.0] - 2026-04-16